refactor: move lib folder to server folder
