+++ src/app/_components/search-input.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { cn } from '@/app/_lib/cn';
+import { cn } from '@/app/(server)/_lib/cn';
 
 type SearchInputProps = {
   value: string;
