+++ src/app/_components/search.tsx
@@ -2,11 +2,11 @@
 import type { FormEvent } from 'react';
 import { useEffect, useState } from 'react';
 
-import { cn } from '@/app/_lib/cn';
+import { cn } from '@/app/(server)/_lib/cn';
 
 import useLocalStorage from '../(client)/_hooks/use-local-storage';
 import { Loader } from '../(server)/_components/ui/loader';
-import { SEARCH_TERM_KEY } from '../_lib/constants';
+import { SEARCH_TERM_KEY } from '../(server)/_lib/constants';
 
 import { SearchInput } from './search-input';
